Connected backend to database

+++ src/main/java/teamnosleep/ebtrescue/EBT/Rescue/repositories/StoreRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import teamnosleep.ebtrescue.EBT.Rescue.data.Store;
 
 @Repository
-public interface StoreRepository extends JpaRepository<Store, Long>, JpaSpecificationExecutor<Store> {
+public interface StoreRepository extends JpaRepository<Store, Integer>, JpaSpecificationExecutor<Store> {
 }
